refactor: migrate AI chat endpoint from Cloudflare Worker to SSR
Replace standalone ai-chat-worker package with server-side rendered endpoint
at pages/api/ai-chat.tsx. This consolidates deployment, improves error handling
with new CF KV bindings, and simplifies the architecture by leveraging Astro's
built-in routing and prerender control.

Changes:
- Replace src/pages/api/ai-chat.ts with pages/api/ai-chat.tsx (SSR endpoint)
- Remove packages/ai-chat-worker/ (standalone CF Worker package, 13 files)
- Remove .github/workflows/ai-chat-worker-deploy.yml (obsolete worker deploy)
- Update wrangler.toml with CF KV bindings (RATE_LIMIT, ANTHROPIC_API_KEY, DOCS_SITE_URL)
- Update docs: ai-assistant.mdx, ai-assistant-api.mdx (new CF bindings table, error codes)
- Update guides/configuration.mdx (fix aiAssistant description)
- Remove reference/ai-chat-worker.mdx (package no longer exists)
