--- .github/workflows/ai-chat-worker-deploy.yml
@@ -1,51 +1,0 @@
-# Deploy the AI Chat Worker to Cloudflare Workers.
-#
-# Triggered on push to `main` when files in packages/ai-chat-worker/ change,
-# or manually via workflow_dispatch.
-
-name: AI Chat Worker Deploy
-
-on:
-  push:
-    branches:
-      - main
-    paths:
-      - 'packages/ai-chat-worker/**'
-  workflow_dispatch:
-
-concurrency:
-  group: ai-chat-worker-deploy
-  cancel-in-progress: false
-
-permissions:
-  contents: read
-
-jobs:
-  deploy:
-    name: Deploy AI Chat Worker
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@93cb6efe18208431cddfb8368fd83d5badbf9bfd # v5
-
-      - name: Setup pnpm
-        uses: pnpm/action-setup@fc06bc1257f339d1d5d8b3a19a8cae5388b55320 # v4
-        with:
-          dest: ~/setup-pnpm-${{ github.job }}
-
-      - name: Setup Node.js
-        uses: actions/setup-node@a0853c24544627f65ddf259abe73b1d18a591444 # v5
-        with:
-          node-version: 22
-
-      - name: Install dependencies
-        run: pnpm install --frozen-lockfile
-
-      - name: Deploy Worker
-        working-directory: packages/ai-chat-worker
-        run: pnpm run deploy
-        env:
-          CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
