--- packages/ai-chat-worker/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@zudo-doc/ai-chat-worker",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "dev": "wrangler dev",
-    "deploy": "wrangler deploy",
-    "typecheck": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20250214.0",
-    "typescript": "^5.9.3",
-    "wrangler": "^4.0.0"
-  }
-}
