--- packages/ai-chat-worker/src/cors.ts
@@ -1,15 +1,0 @@
-const CORS_HEADERS = {
-  "Access-Control-Allow-Origin": "*",
-  "Access-Control-Allow-Methods": "POST, OPTIONS",
-  "Access-Control-Allow-Headers": "Content-Type",
-  "Access-Control-Expose-Headers": "Retry-After",
-  "Access-Control-Max-Age": "86400",
-};
-
-export function corsHeaders(): Record<string, string> {
-  return { ...CORS_HEADERS };
-}
-
-export function handleOptions(): Response {
-  return new Response(null, { status: 204, headers: CORS_HEADERS });
-}
