+++ src/content/docs/reference/ai-assistant-api.mdx
@@ -70,28 +70,32 @@ interface AiChatErrorResponse {
 |--------|-----------|
 | 400 | Invalid JSON body |
 | 400 | `message` is not a non-empty string |
-| 400 | `history` is not an array |
-| 500 | `AI_CHAT_MODE` is not set to `"local"` or `"remote"` |
-| 500 | Claude CLI or Anthropic API call failed |
+| 400 | `message` exceeds 4000 character limit |
+| 400 | Message rejected by input screening (prompt injection guard) |
+| 429 | Rate limit exceeded (includes `Retry-After` header) |
+| 500 | Anthropic API call failed |
 
-## Environment Variables
+## CF Env Bindings
 
-| Variable | Required | Description |
-|----------|----------|-------------|
-| `AI_CHAT_MODE` | Yes | `"local"` or `"remote"` |
-| `ANTHROPIC_API_KEY` | Remote mode only | Anthropic API key |
-| `PUBLIC_ENABLE_MOCKS` | No | Set to `"true"` to enable MSW mock responses (dev mode only, ignored in production) |
+| Binding | Kind | Required | Description |
+|---------|------|----------|-------------|
+| `ANTHROPIC_API_KEY` | secret | Yes | Anthropic API key |
+| `DOCS_SITE_URL` | var | Yes | Deployed docs URL (used to fetch `llms-full.txt`) |
+| `RATE_LIMIT` | KV namespace | Yes | Stores rate-limit counters and audit logs |
+| `RATE_LIMIT_PER_MINUTE` | var | No | Max requests per IP per minute (default `10`) |
+| `RATE_LIMIT_PER_DAY` | var | No | Max requests per IP per day (default `100`) |
+| `PUBLIC_ENABLE_MOCKS` | var | No | Set to `"true"` to enable MSW mock responses (dev mode only) |
 
-## Backend Modes
+## Security
 
-### Local Mode (`AI_CHAT_MODE=local`)
+The endpoint includes layered defenses ported from the legacy standalone worker:
 
-Spawns `claude -p --model haiku` as a subprocess. Documentation context is piped via stdin. 60-second timeout.
-
-### Remote Mode (`AI_CHAT_MODE=remote`)
-
-Calls the Anthropic Messages API with `claude-haiku-4-5-20251001`. Documentation content is included in the system prompt.
+- **Hardened system prompt** — XML-tagged context with explicit guardrails prevents the model from leaking configuration or following off-topic instructions
+- **Input screening** — regex pre-filter rejects common prompt injection patterns before the Claude API is called, without consuming rate limit quota
+- **Rate limiting** — per-IP limits via `RATE_LIMIT` KV (best-effort, fail-open on KV outage)
+- **Audit logging** — every interaction is logged to `RATE_LIMIT` KV with `audit:` prefix (7-day TTL, fire-and-forget)
+- **Message length cap** — messages over 4000 characters are rejected before reaching the API
 
 ## Documentation Context
 
-The endpoint loads `llms-full.txt` (generated by the [llms.txt integration](/docs/guides/llms-txt)) as context. The file is loaded once and cached in memory.
+The endpoint fetches `llms-full.txt` (generated by the [llms.txt integration](/docs/guides/llms-txt)) from `DOCS_SITE_URL` and caches it in memory for the CF Workers isolate lifespan (best-effort, ~1 hour). The content is included in the system prompt as `<documentation>` XML context.
